% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{add_child_nodes}
\alias{add_child_nodes}
\title{Traverse hierarchical code dictionary}
\usage{
add_child_nodes(dfcode, codeVct)
}
\arguments{
\item{dfcode}{hierarchical tree-structured code dictionary with at least columns "coding","selectable","node" and "parent".}

\item{codeVct}{Input code as character vector}
}
\value{
All selectable codes (input code inclusive) as character vector
}
\description{
Given a hierchical dictionary, identify all children codes under the input parent code.
}
\keyword{auxiliary}
