% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{check_dfDefinitions_codes}
\alias{check_dfDefinitions_codes}
\title{Check if the illness/medication codes in definition table are found the coding files}
\usage{
check_dfDefinitions_codes(
  dfDefinitions_processed,
  lst.data.settings,
  code_map_dir = "data/",
  check_expandable_codes_only = T
)
}
\arguments{
\item{dfDefinitions_processed}{definition table as dataframe}

\item{lst.data.settings}{data.setting as dataframe, it includes settings that should be used, e.g. if ICD10 should be looked up case sensitive or not (incase of READ cases are important, dotts should also NOT be interpreted)}

\item{code_map_dir}{the directory storing the coding files}

\item{check_expandable_codes_only}{boolean flag to indicate whether or not to check the only codes to be expanded}
}
\value{
character vector that are missing
}
\description{
Given a definition table and the directory with codings, check if codes in definition table are present in the coding files.
}
\keyword{definition}
