% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{CheckDuplicateTRAITS}
\alias{CheckDuplicateTRAITS}
\title{Check if there is duplicates entries of phenotypes/traits in definition tables}
\usage{
CheckDuplicateTRAITS(df)
}
\description{
This function verify traits are uniquely defined and stop execution otherwise.
}
\keyword{auxiliary}
