% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_cancerregister_to_episodedata.r
\name{convert_cancerregister_to_episodedata}
\alias{convert_cancerregister_to_episodedata}
\title{Convert  illness codes from cancer register columns with corresponding time of diagnosis to episodes format}
\usage{
convert_cancerregister_to_episodedata(
  df,
  field_diagnosis = "40006",
  field_date = "40008",
  field_date_type = "date",
  qc_threshold_year = 0,
  event_code = 1,
  codetype = "numeric"
)
}
\arguments{
\item{df}{dataframe containing the fields}

\item{field_diagnosis}{data field number for illness code, default: 20002}

\item{field_date}{data field number for corresponding time of diagnosis, default: 20008}

\item{field_date_type}{type of time of diagnosis - interpolated_year/interpolated_year/interpolated_year, default:interpolated_year}

\item{qc_threshold_year}{in case of multiple episodes for same illness code, if time difference is larger than qc threshold between the oldest and newest episodes, eventdate will be discarded i.e. set to NA, default:10}

\item{event_code}{code number to indicate if the episodes is a true event 1, self-reported/interpolated event 2 , for which a mean date is taken as event date in case of multiple occurence, or non-event 0 , for which date does not correspond to occurence of illness code. These codes are used in time-to-event analysis, default: 2}

\item{codetype}{data type of the code, either numeric or character, default: numeric}
}
\value{
a data.table object with all episodes
}
\description{
This function takes data fields containing illness codes/time of diagnosis distributed in multiple with each row representing one individual. It processed the data and return in episodes of event for all individuals
}
\examples{
convert_cancerregister_to_episodedata(dfukb,field_diagnosis = "20001",field_date = "20006",qc_treshold_year = 10)
convert_cancerregister_to_episodedata(dfukb,field_diagnosis = "20002",field_date = "20008",qc_treshold_year = 10)
}
\keyword{data}
\keyword{self-reported}
