% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_nurseinterview_to_episodedata.r
\name{convert_col_to_integer}
\alias{convert_col_to_integer}
\title{Convert numbers in character to integer}
\usage{
convert_col_to_integer(col)
}
\value{
character vector of number as integer
}
\description{
This function takes numbers as string and converts into type integer, digits after decimal point is discarded
}
\examples{
convert_col_to_integer("2020")
convert_col_to_integer(c("2047","2020"))
}
\keyword{auxiliary}
