% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{convert_dataframelist_to_lst}
\alias{convert_dataframelist_to_lst}
\title{Helper function to structure the processed definition}
\usage{
convert_dataframelist_to_lst(lst.dfs)
}
\arguments{
\item{lst.dfs}{a list of definition tables (in dataframes)}
}
\value{
restructured list of definition tables organized by dependency status
}
\description{
This function is alternative way to store and output the definition table
}
\keyword{auxiliary}
