% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_nurseinterview_to_episodedata.r
\name{convert_year_to_date}
\alias{convert_year_to_date}
\title{Convert year in decimal unit to date of format yyyy-mm-dd}
\usage{
convert_year_to_date(year)
}
\value{
character vector of date in format yyyy-mm-dd
}
\description{
This function takes numerical year and converts into a date format.
}
\examples{
convert_year_to_date(1999.25)
convert_year_to_date(c(1999.25,2020))
}
\keyword{auxiliary}
