% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{ConvertFactorsToStringReplaceNAInDf}
\alias{ConvertFactorsToStringReplaceNAInDf}
\title{Change factors to strings in a dataframe and assign NA to empty cells}
\usage{
ConvertFactorsToStringReplaceNAInDf(df)
}
\arguments{
\item{df}{dataframe containing the fields}
}
\value{
a data.table object with all episodes
}
\description{
This function takes data fields containing illness codes/time of diagnosis distributed in multiple with each row representing one individual. It processed the data and return in episodes of event for all individuals
}
\examples{
ConvertFactorsToStringReplaceNAInDf(data.frame("a" = c(1,2,3),"p" = c("t","b",""),"y" = c(NA,"w","r")))
}
\keyword{auxiliary}
