% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-ukb-functions.R
\name{default_ukb_fields}
\alias{default_ukb_fields}
\title{Default_ukb_fields}
\usage{
default_ukb_fields()
}
\value{
character vector of data fields
}
\description{
Output the default data fields in a character vector. These data fields are used in converting data to episode data functions and include visit dates,birth ,self reported illness, death.
}
\examples{
default_ukb_fields()
}
