% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_summary_level.r
\name{dotplot_diseases_by_source}
\alias{dotplot_diseases_by_source}
\title{Dotplot by data sources}
\usage{
dotplot_diseases_by_source(
  definitions,
  lst.data,
  df.data.settings,
  vct.identifiers = NULL,
  standardize = TRUE
)
}
\arguments{
\item{definitions}{phenotype/trait specified in definition table (a row in the table)}

\item{lst.data}{list of data table with all episode data}

\item{vct.identifiers}{character vector listing the identifiers in the cohort. This is used to define controls if reference_date is not given}

\item{lst.data.settings}{data frame containing data settings}
}
\value{
gg object
}
\description{
Given phenotypes, make dotplot by data source.
}
\keyword{plot}
