% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{expand_dfDefinitions_processed}
\alias{expand_dfDefinitions_processed}
\title{Expand the illness/medication codes in definition table to include codes that belong to this code}
\usage{
expand_dfDefinitions_processed(
  dfDefinitions_processed,
  lst.data.settings,
  lst.counts
)
}
\arguments{
\item{dfDefinitions_processed}{definition table as dataframe}

\item{lst.data.settings}{data.setting as dataframe, it includes settings that should be used, e.g. if ICD10 should be looked up case sensitive or not (incase of READ cases are important, dotts should also NOT be interpreted)}

\item{lst.counts}{a summary of the raw data produced by \code{get_lst_counts()}}
}
\value{
definition table as dataframe
}
\description{
Explicitly fill in codes covered by the parent codes that are available in the raw data (using lst.counts)stated in the definition table
}
\keyword{definition}
