% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{expand_dfDefinitions_processed2}
\alias{expand_dfDefinitions_processed2}
\title{Expand the illness/medication codes in definition table to include codes that belong to this code}
\usage{
expand_dfDefinitions_processed2(
  dfDefinitions_processed,
  lst.data.settings,
  code_map_dir
)
}
\arguments{
\item{dfDefinitions_processed}{definition table as dataframe}

\item{lst.data.settings}{data.setting as dataframe, it includes settings that should be used, e.g. if ICD10 should be looked up case sensitive or not (incase of READ cases are important, dotts should also NOT be interpreted)}

\item{code_map_dir}{the directory storing the coding files}
}
\value{
definition table as dataframe
}
\description{
Explicitly fill in codes covered by the parent codes stated in the definition table. This differs from \code{expand_dfDefinitions_processed()} in the dictionaries taken - it reads the code dictionaries stated in lst.data.setting and expand to available codes accordingly. Codes not present \strong{will be removed}.
}
\keyword{definition}
