% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{get_allvarnames}
\alias{get_allvarnames}
\title{Retrieve the required data fields in the ukb dataset \emph{ukbxxxxx.tab} as stated in the definition table}
\usage{
get_allvarnames(dfDefinitions_processed, dfhtml = NULL, allow_miss = TRUE)
}
\arguments{
\item{dfDefinitions_processed}{definition table as dataframe}

\item{dfhtml}{the metadata file for main dataset}
}
\value{
a list of character vectors named all_ukb_fields,nondefault_ukb_fields and default_ukb_fields
}
\description{
Survey the main dataset and collect all field names that will be used as specified in definition table. These include basic data fields specified in \code{default_ukb_fields()}, which can be updated accordingly, and the fields in touchscreen columns. If metadata is supplied , it will check if these fields are present in the maindataset.
}
\examples{
dfDefinitions_processed <- ProcessDfDefinitions(fread("definitions.tsv", colClasses = 'character', data.table = FALSE))
dfhtml <- read_ukb_metadata(fhtml)
dfDefinitions_ukb_fields <- get_allvarnames(dfDefinitions_processed,dfhtml)
}
\keyword{definition}
