% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_and_visualize_by_source.r
\name{get_case_count_by_source}
\alias{get_case_count_by_source}
\title{Get case-control status by data sources}
\usage{
get_case_count_by_source(
  definition,
  lst.data,
  df.data.settings,
  df_reference_dates = NULL,
  vct_identifiers = NULL,
  standardize = TRUE
)
}
\arguments{
\item{lst.data}{list of data table with all episode data}

\item{definitions}{phenotype/trait specified in definition table (a row in the table)}

\item{lst.data.settings}{data frame containing data settings}

\item{reference_date}{reference dates for each individuals in the whole cohort as a named vector}
}
\value{
a data table : case control status according to different sources and one column based on any of the sources.
}
\description{
Given a phenotype, a list of episode data and reference dates per individual, identify cases and control status as of *reference date" by source.
}
\examples{
get_case_count_by_source(cancer_source,definition=dfDefinitions_processed_expanded \%>\% filter(TRAIT=="Nicm"), lst.data,lst.data.settings,lst.identifiers=dfukb$f.eid))
}
\keyword{time-to-event}
