% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_event_tables.r
\name{get_cases}
\alias{get_cases}
\title{Get case for a phenotype}
\usage{
get_cases(
  definitions,
  lst.data,
  df.data.settings,
  df_reference_dt = NULL,
  include_secondary_recurrence = FALSE,
  verbose = TRUE,
  window_fu_days_mask = 0,
  ...
)
}
\arguments{
\item{definitions}{phenotype/trait specified in definition table (a row in the table)}

\item{lst.data}{list of data table with all episode data}

\item{df.data.settings}{data frame containing data settings}

\item{window_fu_days_mask}{number of days that future events should not be counted. e.g. you could only count events after 10 days from the reference visit to avoid events related to the reference date.}

\item{reference_date}{reference dates for each individuals in the whole cohort as a named vector}
}
\value{
a list of 2 data tables : all events for valid cases and an event summary containing time to event information for these individuals.
}
\description{
Given a phenotype, a list of episode data and reference dates per individual, identify cases /exclude cases fulfilling exclusion criteria and compute the time to event data for these individuals.  If no reference date is given then the date of first available event will be taken as reference date for each individual.
}
\examples{
get_cases(definitions=dfDefinitions_processed_expanded \%>\% filter(TRAIT=="Nicm"), lst.data,df.data.settings, reference_date=setNames(as.Date(as.character(dfukb$f.53.0.0),format="\%Y-\%m-\%d"),dfukb$identifier))
}
\keyword{time-to-event}
