% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-ukb-functions.R
\name{get_lst_counts}
\alias{get_lst_counts}
\title{Get summary counts on a list of data tables}
\usage{
get_lst_counts(lst.data, lst.data.settings)
}
\arguments{
\item{lst.data}{list of data tables}

\item{lst.data.settings}{data.setting as dataframe}
}
\value{
a list of data tables of summary counts organized in classifications
}
\description{
This function reads list of episode data in data tables and return summary counts by classification systems according to lst.data.settings.
}
\examples{
lst.data.settings <-fread(data.settings.tsv)
lst.data <- list()
lst.data$ts <- convert_touchscreen_to_episodedata(dfukb,ts_conditions = dfDefinitions_processed$TS)
lst.data$tte.sr.20002 <- convert_nurseinterview_to_episodedata(dfukb,field_sr_diagnosis = "20002",field_sr_date = "20008",qc_treshold_year = 10)
lst.data <- append(lst.data,read_hesin_data(fhesin ,fhesin_diag ,fhesin_oper ))
get_lst_counts(lst.data,lst.data.settings)
}
