% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_event_tables.r
\name{get_survival_data}
\alias{get_survival_data}
\title{Get survival data}
\usage{
get_survival_data(
  def,
  lst.data,
  lst.data.settings,
  include_secondary_recurrence = FALSE,
  window_days_mask = 0
)
}
\arguments{
\item{def}{phenotype/trait specified in definition table (a row in the table)}

\item{lst.data}{list of data table with all episode data}

\item{lst.data.settings}{data frame containing data settings}

\item{window_days_mask}{number of days that future events should not be counted since first diagnosis.Relevant when the death record is the only record i.e. t=0}
}
\value{
a data table summarizing time to event information for these individuals.
}
\description{
Given a phenotype and a list of episode data , compute the survival time after first diagnoisis in cases only
}
\examples{
get_survival_data(dfDefinitions_processed_expanded[14,],lst.data,lst.data.settings)
}
\keyword{time-to-event}
