% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ukb_data.R
\name{isNullNaNan}
\alias{isNullNaNan}
\title{Check if string variable is valid}
\usage{
isNullNaNan(x)
}
\value{
logical TRUE / FALSE
}
\description{
Any of these will cause file.exists() to throw an error
}
