% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{lookup.codes}
\alias{lookup.codes}
\title{Helper function to look up the dependencies}
\usage{
lookup.codes(df, lookupcolumn = "Exclude_from_cases", VctAllColumns)
}
\arguments{
\item{df}{definition table as dataframe}

\item{lookupcolumn}{name of dependency column that needs lookup}

\item{VctAllColumns}{names of columns in which codes should be updated}
}
\value{
a new dataframe containing definitions as the result of the lookup
}
\description{
This function is used for trait dependency lookup
}
\keyword{auxiliary}
