% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_summary_level.r
\name{make_upsetplot}
\alias{make_upsetplot}
\title{Make upset plot}
\usage{
make_upsetplot(definition, lst.data, df.data.settings, df.reference.dates)
}
\arguments{
\item{lst.data}{list of data table with all episode data}

\item{definitions}{phenotype/trait specified in definition table (a row in the table)}

\item{lst.data.settings}{data frame containing data settings}

\item{reference_date}{reference dates for each individuals in the whole cohort as a named vector}
}
\value{
upset plot
}
\description{
Given a phenotype and reference date , make an upset plot of cases (Hx) by the reference dates by source.
}
\examples{
make_upsetplot(definition=dfDefinitions_processed_expanded\%>\%filter(TRAIT==trait),lst.harmonized.data$lst.data,df.data.settings,df.reference.dates = df_reference_dt_v0)
}
\keyword{upset}
