% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-ukb-functions.R
\name{parse_gp_clinical_values}
\alias{parse_gp_clinical_values}
\title{Parse the value columns of the Primary Care clinical records}
\usage{
parse_gp_clinical_values(fgp)
}
\arguments{
\item{fgp}{Path to GP clinical event records}
}
\value{
list of data.table objects with all episodes
}
\description{
This function parse the free text fields of the GP clinical records. Information contains in the columns include test results and remarks
}
\examples{
parse_gp_clinical_values("gpclinical.txt" )
}
