% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{parseIncludeExcludeCol}
\alias{parseIncludeExcludeCol}
\title{Helper function to parse the trait dependencies}
\usage{
parseIncludeExcludeCol(df, InExCol, concat_to_df = FALSE, VctAllColumns)
}
\arguments{
\item{df}{definition table as dataframe}

\item{InExCol}{name of dependency column that needs lookup}

\item{concat_to_df}{option to update the input df otherwise store the codes in a new df,default: TRUE}

\item{VctAllColumns}{names of columns in which codes should be updated}
}
\value{
the original dataframe or a new dataframe with codes
}
\description{
Each composite trait with dependencies considered a tree as chain of dependency is possible. To parse and retrieve all codes needed, this function first searches for a leaf nodes and then add all codes of nodes along the way until root is reached; this operation is done recursively until all dependent codes are filled into the df
}
\keyword{definition}
