% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{pasteRemoveNA}
\alias{pasteRemoveNA}
\title{Custom paste function with formatting step to remove redundant spaces}
\usage{
pasteRemoveNA(..., sep = " ", collapse = NULL, na.rm = F)
}
\value{
formatted object
}
\description{
This function is used for formatting the character vectors
}
\keyword{auxiliary}
