% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_summary_level.r
\name{plot_disease_timeline_by_source}
\alias{plot_disease_timeline_by_source}
\title{data source trajectory over time}
\usage{
plot_disease_timeline_by_source(
  definition,
  lst.data,
  df.data.settings,
  vct.identifiers
)
}
\arguments{
\item{lst.data}{list of data table with all episode data}

\item{df.data.settings}{data frame containing data settings}

\item{vct.identifiers}{character vector listing the identifiers in the cohort. This is used to define controls if reference_date is not given}

\item{definitions}{phenotype/trait specified in definition table (a row in the table)}
}
\value{
ggplot object
}
\description{
Given a phenotype definition and list of data-tables with events stratified by sources, plot the case growth over time by sources
}
\keyword{event}
\keyword{phenotype}
\keyword{stats}
\keyword{timeline}
