% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessdfDefinitions.R
\name{PreProcessDfDefinitions}
\alias{PreProcessDfDefinitions}
\title{Preprocess the definition table}
\usage{
PreProcessDfDefinitions(df, VctAllColumns, VctColstoupper = NULL)
}
\arguments{
\item{df}{definitiion table as dataframe}

\item{VctAllColumns}{character vector containing all column names of interest (as shown in table) where preprocessing is performed}

\item{VctColstoupper}{character vector containing all column names without brackets for which the characters in columns will be converted to uppercase.}
}
\value{
df
}
\description{
This function preprocess definition table - formatting and remove extra characters as well as additional description following column names (in brackets)
}
\examples{
PreProcessDfDefinitions(df,c("TS(Touchscreen)","ICD10", "ICD9", "OPCS4","OPCS3","READ2","READ2_drugs", "CTV3","BNF","DMD","f.20001(sr_cancer)", "f.20002(sr_noncancer)", "f.20003(sr_med)", "f.20004(sr_oper)"),c("ICD10","ICD9","OPCS4","OPCS3"))
}
\keyword{definition}
