% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_summary_level.r
\name{rand_col_vec}
\alias{rand_col_vec}
\title{return random colors of specified length}
\usage{
rand_col_vec(vec_leng = 5, alpha = 0.75, cust_seed = NULL)
}
\arguments{
\item{vec_leng}{phenotype/trait specified in definition table (a row in the table)}

\item{alpha}{list of data table with all episode data}

\item{cust_seed}{data frame containing data settings}
}
\value{
vector of rgb colour codes
}
\description{
Given a phenotype, a list of episode data and reference dates per individual, identify cases and control status as of *reference date" by source.
}
