% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-ukb-functions.R
\name{read_death_data}
\alias{read_death_data}
\title{Read Mortality data}
\usage{
read_death_data(fdeath_portal, fdeath_cause_portal)
}
\arguments{
\item{fdeath_portal}{Path to file with DEATH table}

\item{fdeath_cause_portal}{Path to file with DEATH_CAUSE table}
}
\value{
a data.table object with all episodes
}
\description{
This function reads death data on the Data Portal.
}
\examples{
read_death_data("death.txt","death_cause.txt" )
}
