% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ukb_data.R
\name{read_definition_table}
\alias{read_definition_table}
\title{Process definition table}
\usage{
read_definition_table(f.definition, f.data.setting, dir.code.map)
}
\arguments{
\item{f.definition}{path to definition table (.tsv)}

\item{dir.code.map}{path to directory of the code mappings}

\item{f.data.settings}{data frame containing data settings}
}
\value{
expanded definition table
}
\description{
Given a definition table, check, clean, format as well as to expand the codes.
}
\examples{
dfDefinitions_processed_expanded<-read_defnition_table(fdefinitions,fdata_setting,dir.code.map=paste0(repo_dir,"inst/extdata/"))
}
\keyword{definitions}
