% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-ukb-functions.R
\name{read_gp_clinical_data}
\alias{read_gp_clinical_data}
\title{Read Primary Care clinical event records}
\usage{
read_gp_clinical_data(fgp)
}
\arguments{
\item{fgp}{Path to GP clinical event records}
}
\value{
list of data.table objects with all episodes
}
\description{
This function reads the record-level clinical events from General Practitioner (GP) data. Refer to official UKB documetation for more information regarding this data.
}
\examples{
read_gp_clinical_data("gpclinical.txt" )
}
