% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-ukb-functions.R
\name{read_hesin_data}
\alias{read_hesin_data}
\title{Read Hospital Inpatient Data}
\usage{
read_hesin_data(fhesin, fhesin_diag, fhesin_oper, add_extra_hesin_columns = F)
}
\arguments{
\item{fhesin}{Path to HESIN (master file)}

\item{fhesin_diag}{Path to HESIN_DIAG file containing diagnosis codes}

\item{fhesin_oper}{Path to HESIN_OPER file containing Operations and procedural codes}

\item{add_extra_hesin_columns}{if True, adds extra columns "ins_index","source"}
}
\value{
list of data.table objects with all episodes
}
\description{
This function reads the record-level HospitalEpisodeStatistics (HES) data organized in several data tables
}
\examples{
read_hesin_data("hesin.txt" ,"hesin_diag.txt" ,"hesin_oper.txt" )
}
