% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-ukb-functions.R
\name{read_ukb_metadata}
\alias{read_ukb_metadata}
\title{Read the data dictionary for main dataset}
\usage{
read_ukb_metadata(fhtml)
}
\arguments{
\item{fhtml}{Path to file}
}
\value{
metadata for main dataset as dataframe
}
\description{
The metadata file \emph{ukbxxxxx.html} can be generated using ukb utility. This function reads the html formatted file and return metadata in a dataframe.
}
\examples{
read_ukb_metadata ("ukb12345.html")
}
