% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-ukb-functions.R
\name{read_ukb_tabdata}
\alias{read_ukb_tabdata}
\title{Read the main dataset}
\usage{
read_ukb_tabdata(fukb, dfhtml, fields_to_keep = default_ukb_fields())
}
\arguments{
\item{fukb}{Path to the main dataset (.tab) file}

\item{dfhtml}{dataframe containing the metadata, can be obtained by \code{read_ukb_metadata()}.}

\item{fields_to_keep}{the datafields to keep as a character vector, default: default_ukb_fields()}
}
\value{
main dataset as dataframe with only selected data fields
}
\description{
This function reads the dataset \emph{ukbxxxxx.tab} generated using ukb utility and keep only data fields specified. Depending on the size of the dataset the process may take some time.
}
\examples{
fukbtab <- paste(pheno_dir,"ukb12345.tab",sep="")
dfhtml<- read_ukb_metadata ("ukb12345.html")
dfukb <- read_ukb_tabdata(fukbtab,dfhtml,fields_to_keep = default_ukb_fields())
}
