% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{:=}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:assign]{:=}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

