% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-ukb-functions.R
\name{ReplaceNAWithNearestNonNAOnTheLeft}
\alias{ReplaceNAWithNearestNonNAOnTheLeft}
\title{Replace NA With Nearest Non-NA On The Left}
\usage{
ReplaceNAWithNearestNonNAOnTheLeft(Vct)
}
\value{
imputed character vector
}
\description{
Used in parsing meta data file ukbxxxxx.html.
}
