% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_summary_level.r
\name{get_case_count_by_source}
\alias{get_case_count_by_source}
\title{Get case-control status (Any) by data sources}
\usage{
get_case_count_by_source(
  definition,
  lst.data,
  df.data.settings,
  vct.identifiers,
  remove_event0 = TRUE
)
}
\arguments{
\item{definition}{phenotype/trait specified in definition table (a row in the table)}

\item{lst.data}{list of data table with all episode data}

\item{df.data.settings}{data frame containing data settings}

\item{vct.identifiers}{character vector listing the identifiers in the cohort. This is used to define controls if reference_date is not given}
}
\description{
Given a phenotype, a list of episode data and reference dates per individual, identify cases and control status as of *reference date" by source.
}
\keyword{event}
\keyword{phenotype}
\keyword{stats}
