% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_individual_timeline}
\alias{plot_individual_timeline}
\title{Get all episodes for a phenotype}
\usage{
plot_individual_timeline(
  df.data.settings,
  ind_all_event_dt = NULL,
  lst.data = NULL,
  ind_identifier = 1234567,
  plot_medication = FALSE,
  keep_classifications = NULL,
  font_size = 5
)
}
\arguments{
\item{df.data.settings}{data frame containing data settings}

\item{ind_all_event_dt}{a data table containing all events for the target individual}

\item{lst.data}{list of data tables with all episode data collapsed to separate datatables}

\item{ind_identifier}{identifier of the target individual for the timeline to be visualized}
}
\value{
a data table with all events
}
\description{
Given a phenotype and a list of episode data , extract events for this phenotype from all data sources
}
\examples{
lst.data.identifier<-lapply(lst.data,function(x) {x[, ("identifier") := lapply(.SD, as.numeric), .SDcols = "identifier"] })  set eid to numeric
lst.data.identifier<-lapply(lst.data,function(x) {setkey(x,identifier) }) # double check that everything has the same setkey.
plot_individual_timeline(df.data.settings,NULL,lst.data.identifier,ind_identifier="1234567")
}
\keyword{time-to-event}
