% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data-ukb-functions.R
\name{read_gp_script_data}
\alias{read_gp_script_data}
\title{Read Primary Care prescription records}
\usage{
read_gp_script_data(fgp, add_extra_script_columns = F)
}
\arguments{
\item{fgp}{Path to GP prescription records}

\item{add_extra_script_columns}{whether or not to keep drug_name and quantity columns, default: False}
}
\value{
list of data.table objects with all episodes
}
\description{
This function reads the prescription records from General Practitioner (GP) data. Refer to official UKB documetation for more information regarding this data.
}
\examples{
read_gp_clinical_data("gpscripts.txt" )
}
